+++ app/src/main/java/com/example/emlody/Activities/AnalyzeEmotionActivity.java
@@ -1,19 +1,12 @@
 package com.example.emlody.Activities;
 
 import android.Manifest;
-import android.app.Dialog;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
 import android.widget.ImageView;
-import android.widget.ListView;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.activity.result.ActivityResultLauncher;
@@ -33,8 +26,6 @@ import com.google.gson.Gson;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
@@ -59,23 +50,6 @@ public class AnalyzeEmotionActivity extends AppCompatActivity {
     FloatingActionButton galleryFloatingActionButton;
     FloatingActionButton cameraFloatingActionButton;
 
-    Dialog playlistsDialog;
-
-    Dialog emotionsDialog;
-
-    TextView goBackText, tellText, doneText;
-
-    String[] emotions = {"Happy", "Sad", "Angry", "Exited", "Nervous", "Fear"};
-    ArrayAdapter<String> adapter;
-
-    List<String> chosenEmotionsByUser;
-
-    ListView emotionsListView;
-
-    ArrayList emotionsArrayList;
-
-
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -92,6 +66,7 @@ public class AnalyzeEmotionActivity extends AppCompatActivity {
             iGallery.setData(MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
             imageLauncher.launch("image/*");
         });
+
     }
 
 private Uri createUri(){
@@ -225,63 +200,6 @@ private void checkCameraPermissionsAndOpenCamera(){
         startActivity(intent);
     }
 
-    public void showPlayListsDialog() {
-
-        playlistsDialog = new Dialog(AnalyzeEmotionActivity.this);
-        playlistsDialog.setContentView(R.layout.playlists_dialog_layout);
-        playlistsDialog.getWindow().setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        playlistsDialog.setCancelable(false);
-        playlistsDialog.getWindow().getAttributes().windowAnimations = R.style.animation;
-        tellText = findViewById(R.id.okay_text);
-        goBackText = findViewById(R.id.cancel_text);
-        goBackText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                playlistsDialog.dismiss();
-            }
-        });
-
-        tellText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                playlistsDialog.dismiss();
-                showEmotionsDialog();
-            }
-        });
-    }
-
-    public void showEmotionsDialog() {
-        chosenEmotionsByUser = new ArrayList<>();
-        emotionsDialog = new Dialog(AnalyzeEmotionActivity.this);
-        emotionsDialog.setContentView(R.layout.emotions_dialog_layout);
-        emotionsDialog.getWindow().setLayout(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        emotionsDialog.setCancelable(false);
-        emotionsDialog.getWindow().getAttributes().windowAnimations = R.style.animation;
-        emotionsListView = findViewById(R.id.moods_listview);
-        emotionsArrayList = new ArrayList();
-        adapter = new ArrayAdapter<>
-                (AnalyzeEmotionActivity.this,
-                        android.R.layout.select_dialog_multichoice,
-                        emotionsArrayList
-                );
-        emotionsListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                chosenEmotionsByUser.add(adapter.getItem(position));
-            }
-        });
-        this.emotionsListView.setAdapter(adapter);
-        doneText = findViewById(R.id.done_text);
-        doneText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                emotionsDialog.dismiss();
-                //todo call the server with emotions chosen by user
-            }
-        });
-
-        emotionsDialog.show();
-    }
 }
