+++ app/src/main/res/layout/playlists_dialog_layout.xml
@@ -15,7 +15,7 @@
         android:src="@drawable/questionmark" />
 
     <TextView
-        android:id="@+id/textview"
+        android:id="@+id/information"
         android:layout_width="wrap_content"
         android:layout_height="232dp"
         android:layout_below="@id/imageView"
@@ -23,13 +23,13 @@
         android:padding="10dp"
         android:text="We apologize, it seems we were not able to determine how you are feeling right now. Would you like to go back and try again or tell us how are you currently feeling?"
         android:textAlignment="center"
-        android:textSize="25sp"
+        android:textSize="20sp"
         android:textStyle="bold" />
 
     <RelativeLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_below="@id/textview"
+        android:layout_below="@id/information"
         android:layout_margin="15dp"
         android:layout_centerHorizontal="true">
 
@@ -41,7 +41,7 @@
             android:text="Tell"
             android:textStyle="bold"
             android:textColor="#3F51B5"
-            android:textSize="30sp" />
+            android:textSize="20sp" />
 
         <TextView
             android:id="@+id/cancel_text"
@@ -53,7 +53,7 @@
             android:text="Try Again"
             android:textStyle="bold"
             android:textColor="#3F51B5"
-            android:textSize="30sp" />
+            android:textSize="20sp" />
 
     </RelativeLayout>
